+++ Source/BugsnagEvent.m
@@ -221,6 +221,11 @@ static NSString *const DEFAULT_EXCEPTION_TYPE = @"cocoa";
 @property(readonly, strong, nullable) BugsnagBreadcrumbs *breadcrumbs;
 @end
 
+@interface BugsnagSession ()
+@property NSUInteger unhandledCount;
+@property NSUInteger handledCount;
+@end
+
 @interface BugsnagEvent ()
 
 /**
@@ -623,7 +628,6 @@ initWithErrorName:(NSString *_Nonnull)name
         }
         _overrides = metadata;
     }
-    
 }
 
 - (NSDictionary *)toJson {