+++ Tests/BugsnagTests.m
@@ -8,6 +8,7 @@
 // Unit tests of global Bugsnag behaviour
 
 #import "Bugsnag.h"
+#import "BugsnagClient.h"
 #import "BugsnagTestConstants.h"
 #import <XCTest/XCTest.h>
 
@@ -20,6 +21,10 @@
 @interface BugsnagConfiguration ()
 @property(nonatomic, readwrite, strong) NSMutableArray *onSendBlocks;
 @property(readwrite, retain, nullable) BugsnagMetadata *metadata;
+@end
+
+@interface BugsnagClient ()
+@property (nonatomic, strong) NSString *lastOrientation;
 @end
 
 @interface BugsnagTests : XCTestCase
@@ -386,7 +391,7 @@
     [Bugsnag notify:exception1];
     
     // Both called?
-    [self waitForExpectations:@[expectation1, expectation2] timeout:5.0];
+    [self waitForExpectations:@[expectation1, expectation2] timeout:10.0];
     
     [Bugsnag removeOnSendBlock:block2];
     XCTAssertEqual([[[Bugsnag configuration] onSendBlocks] count], 1);
@@ -396,9 +401,32 @@
     NSException *exception2 = [[NSException alloc] initWithName:@"exception1" reason:@"reason1" userInfo:nil];
     [Bugsnag notify:exception2];
     // One removed, should only call one
-    [self waitForExpectations:@[expectation3, expectation4] timeout:5.0];
+    [self waitForExpectations:@[expectation3, expectation4] timeout:10.0];
 
     [self waitForExpectations:@[expectation5, expectation6] timeout:1.0];
 }
+
+/**
+ * Test that the Orientation -> string mapping is as expected
+ * NOTE: should be moved to BugsnagClientTests when that file exists
+ */
+#if TARGET_OS_IOS
+NSString *BSGOrientationNameFromEnum(UIDeviceOrientation deviceOrientation);
+- (void)testBSGOrientationNameFromEnum {
+    XCTAssertEqualObjects(BSGOrientationNameFromEnum(UIDeviceOrientationPortraitUpsideDown), @"portraitupsidedown");
+    XCTAssertEqualObjects(BSGOrientationNameFromEnum(UIDeviceOrientationPortrait), @"portrait");
+    XCTAssertEqualObjects(BSGOrientationNameFromEnum(UIDeviceOrientationLandscapeRight), @"landscaperight");
+    XCTAssertEqualObjects(BSGOrientationNameFromEnum(UIDeviceOrientationLandscapeLeft), @"landscapeleft");
+    XCTAssertEqualObjects(BSGOrientationNameFromEnum(UIDeviceOrientationFaceUp), @"faceup");
+    XCTAssertEqualObjects(BSGOrientationNameFromEnum(UIDeviceOrientationFaceDown), @"facedown");
+    
+    XCTAssertNil(BSGOrientationNameFromEnum(-1));
+    XCTAssertNil(BSGOrientationNameFromEnum(99));
+    
+    BugsnagClient *client = [BugsnagClient new];
+    [client setLastOrientation:@"testOrientation"];
+    XCTAssertEqualObjects([client lastOrientation], @"testOrientation");
+}
+#endif
 
 @end